requirements.txt
setuptools
websockets
hackerforms
abstra